Added ability to manually shift text horizontally and/or vertically using new builder method: textPostionShift(int textXPosShift, int textYPosShift)
Shifts from default position, rather than setting a new absolute
position.
