--- ShowcaseViewAndfaulknerMod/res/drawable/button_normal.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright 2014 Alex Curran
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item>
-        <shape android:shape="rectangle">
-            <padding android:bottom="2dp" />
-            <solid android:color="#99FFFFFF" />
-        </shape>
-    </item>
-
-    <!--<item>-->
-        <!--<shape android:shape="rectangle">-->
-            <!--<padding android:top="2dp" />-->
-            <!--<solid android:color="#33FFFFFF" />-->
-        <!--</shape>-->
-    <!--</item>-->
-
-    <item>
-        <shape android:shape="rectangle">
-            <solid android:color="#FFFFFF" />
-        </shape>
-    </item>
-
-</layer-list>
