--- ShowcaseViewAndfaulknerMod/res/values/dimens.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright 2014 Alex Curran
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<resources>
-    <dimen name="button_margin">12dp</dimen>
-    <dimen name="text_padding">24dp</dimen>
-    <dimen name="action_bar_offset">66dp</dimen>
-    <dimen name="showcase_radius">94dp</dimen>
-    <dimen name="showcase_radius_inner">96dp</dimen>
-    <dimen name="showcase_radius_outer">128dp</dimen>
-</resources>
