--- ShowcaseViewAndfaulknerMod/res/values/styles.xml
@@ -1,61 +1,0 @@
-<!--
-  ~ Copyright 2014 Alex Curran
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<resources xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <style name="ShowcaseButton">
-        <item name="android:layout_width">wrap_content</item>
-        <item name="android:layout_height">wrap_content</item>
-        <item name="android:minWidth">96dp</item>
-        <item name="android:gravity">center</item>
-        <item name="android:textStyle">bold</item>
-        <item name="android:textColor">#FFFFFF</item>
-        <item name="android:background">@drawable/button</item>
-    </style>
-
-    <style name="ShowcaseView.Light">
-        <item name="sv_titleTextAppearance">@style/TextAppearance.ShowcaseView.Title</item>
-        <item name="sv_detailTextAppearance">@style/TextAppearance.ShowcaseView.Detail.Light</item>
-        <item name="sv_backgroundColor">#3333B5E5</item>
-        <item name="sv_buttonText">@android:string/ok</item>
-    </style>
-
-    <style name="ShowcaseView">
-        <item name="sv_titleTextAppearance">@style/TextAppearance.ShowcaseView.Title</item>
-        <item name="sv_detailTextAppearance">@style/TextAppearance.ShowcaseView.Detail</item>
-        <item name="sv_backgroundColor">#3333B5E5</item>
-        <item name="sv_buttonText">@android:string/ok</item>
-    </style>
-
-    <style name="TextAppearance.ShowcaseView.Title" parent="android:style/TextAppearance.Large">
-        <item name="android:textColor">#33B5E5</item>
-    </style>
-
-    <style name="TextAppearance.ShowcaseView.Detail" parent="android:style/TextAppearance">
-        <item name="android:textColor">#FFFFFF</item>
-    </style>
-
-    <style name="TextAppearance.ShowcaseView.Detail.Light">
-        <item name="android:textColor">#333333</item>
-        <item name="android:shadowColor">#FF0000</item>
-        <item name="android:shadowRadius">0</item>
-    </style>
-
-    <style name="TextAppearance.ShowcaseView.Title.Light">
-        <item name="android:shadowRadius">0</item>
-    </style>
-
-</resources>
