--- ShowcaseViewAndfaulknerMod/src/com/github/amlcurran/showcaseview/ApiUtils.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2014 Alex Curran
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.amlcurran.showcaseview;
-
-import android.annotation.TargetApi;
-import android.os.Build;
-import android.view.View;
-
-public class ApiUtils {
-
-    public boolean isCompatWith(int versionCode) {
-        return Build.VERSION.SDK_INT >= versionCode;
-    }
-
-    public boolean isCompatWithHoneycomb() {
-        return isCompatWith(Build.VERSION_CODES.HONEYCOMB);
-    }
-
-    @TargetApi(14)
-    public void setFitsSystemWindowsCompat(View view) {
-        if (isCompatWith(Build.VERSION_CODES.ICE_CREAM_SANDWICH)) {
-            view.setFitsSystemWindows(true);
-        }
-    }
-}
