--- ShowcaseViewAndfaulknerMod/src/com/github/amlcurran/showcaseview/Calculator.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2014 Alex Curran
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.amlcurran.showcaseview;
-
-import android.graphics.Point;
-import android.view.View;
-
-/**
- * Calculates various items for use with ShowcaseView
- */
-class Calculator {
-
-    static Point getShowcasePointFromView(View view) {
-        Point result = new Point();
-        result.x = view.getLeft() + view.getWidth() / 2;
-        result.y = view.getTop() + view.getHeight() / 2;
-        return result;
-    }
-
-}
