--- ShowcaseViewAndfaulknerMod/src/com/github/amlcurran/showcaseview/targets/AppCompatReflector.java
@@ -1,59 +1,0 @@
-/*
- * Copyright 2014 Alex Curran
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.amlcurran.showcaseview.targets;
-
-import android.app.Activity;
-import android.view.View;
-import android.view.ViewParent;
-
-/**
- * Created by Alex on 27/10/13.
- */
-class AppCompatReflector implements Reflector {
-
-    private Activity mActivity;
-
-    public AppCompatReflector(Activity activity) {
-        mActivity = activity;
-    }
-
-    @Override
-    public ViewParent getActionBarView() {
-        return getHomeButton().getParent().getParent();
-    }
-
-    @Override
-    public View getHomeButton() {
-        View homeButton = mActivity.findViewById(android.R.id.home);
-        if (homeButton != null) {
-            return homeButton;
-        }
-        int homeId = mActivity.getResources().getIdentifier("home", "id", mActivity.getPackageName());
-        homeButton = mActivity.findViewById(homeId);
-        if (homeButton == null) {
-            throw new RuntimeException(
-                    "insertShowcaseViewWithType cannot be used when the theme " +
-                            "has no ActionBar");
-        }
-        return homeButton;
-    }
-
-    @Override
-    public void showcaseActionItem(int itemId) {
-
-    }
-}
