--- ShowcaseViewAndfaulknerMod/src/com/github/amlcurran/showcaseview/targets/Reflector.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2014 Alex Curran
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.amlcurran.showcaseview.targets;
-
-import android.view.View;
-import android.view.ViewParent;
-
-interface Reflector {
-    View getHomeButton();
-
-    void showcaseActionItem(int itemId);
-
-    ViewParent getActionBarView();
-
-    public enum ActionBarType {
-        STANDARD, APP_COMPAT, ACTIONBAR_SHERLOCK
-    }
-}
